move needs review rule list to get-rule-inclusions for importing
